+++ src/consts.rs
@@ -1,0 +1,4 @@
+// Service identifier
+const SERVICE_ID: u16 = 1;
+// Starting balance of a newly created wallet
+const INIT_BALANCE: u64 = 100;
