+++ src/currency/schema.rs
@@ -1,0 +1,47 @@
+use exonum::storage::{Fork, MapIndex, Snapshot};
+use exonum::crypto::{PublicKey};
+
+encoding_struct! {
+    struct Wallet {
+        pub_key: &PublicKey,
+        name: &str,
+        balance: u64,
+    }
+}
+
+impl Wallet {
+    pub fn increase(self, amount: u64) -> Self {
+        let balance = self.balance() + amount;
+        Self::new(self.pub_key(), self.name(), balance)
+    }
+
+    pub fn decrease(self, amount: u64) -> Self {
+        let balance = self.balance() - amount;
+        Self::new(self.pub_key(), self.name(), balance)
+    }
+}
+
+pub struct CurrencySchema<T> {
+    view: T,
+}
+
+impl<T: AsRef<Snapshot>> CurrencySchema<T> {
+    pub fn new(view: T) -> Self {
+        CurrencySchema { view }
+    }
+
+    pub fn wallets(&self) -> MapIndex<&Snapshot, PublicKey, Wallet> {
+        MapIndex::new("cryptocurrency.wallets", self.view.as_ref())
+    }
+
+    // Utility method to quickly get a separate wallet from the storage
+    pub fn wallet(&self, pub_key: &PublicKey) -> Option<Wallet> {
+        self.wallets().get(pub_key)
+    }
+}
+
+impl<'a> CurrencySchema<&'a mut Fork> {
+    pub fn wallets_mut(&mut self) -> MapIndex<&mut Fork, PublicKey, Wallet> {
+        MapIndex::new("cryptocurrency.wallets", &mut self.view)
+    }
+}
