+++ src/lib.rs
@@ -7,164 +7,24 @@ extern crate serde;
 extern crate serde_derive;
 extern crate serde_json;
 
-use exonum::api::{ServiceApiState, ServiceApiBuilder, self};
-use exonum::blockchain::{Blockchain, ExecutionError,
-                         ExecutionResult, Service, Transaction,
-                         TransactionSet};
-use exonum::crypto::{Hash, PublicKey};
-use exonum::encoding;
-use exonum::encoding::serialize::FromHex;
-use exonum::messages::{Message, RawTransaction};
-use exonum::node::{ApiSender, TransactionSend};
-use exonum::storage::{Fork, MapIndex, Snapshot};
+use exonum::{
+    api::{ServiceApiState, ServiceApiBuilder, self},
+    blockchain::{Service, Transaction, TransactionSet},
+    crypto::{Hash, PublicKey},
+    encoding,
+    messages::{RawTransaction},
+    node::{TransactionSend},
+    storage::{Snapshot},
+};
+
+mod currency;
+use currency::{
+    schema::{Wallet, CurrencySchema},
+    transactions::CurrencyTransactions
+};
 
 // Service identifier
 const SERVICE_ID: u16 = 1;
-// Starting balance of a newly created wallet
-const INIT_BALANCE: u64 = 100;
-
-encoding_struct! {
-    struct Wallet {
-        pub_key: &PublicKey,
-        name: &str,
-        balance: u64,
-    }
-}
-
-impl Wallet {
-    pub fn increase(self, amount: u64) -> Self {
-        let balance = self.balance() + amount;
-        Self::new(self.pub_key(), self.name(), balance)
-    }
-
-    pub fn decrease(self, amount: u64) -> Self {
-        let balance = self.balance() - amount;
-        Self::new(self.pub_key(), self.name(), balance)
-    }
-}
-
-pub struct CurrencySchema<T> {
-    view: T,
-}
-
-impl<T: AsRef<Snapshot>> CurrencySchema<T> {
-    pub fn new(view: T) -> Self {
-        CurrencySchema { view }
-    }
-
-    pub fn wallets(&self) -> MapIndex<&Snapshot, PublicKey, Wallet> {
-        MapIndex::new("cryptocurrency.wallets", self.view.as_ref())
-    }
-
-    // Utility method to quickly get a separate wallet from the storage
-    pub fn wallet(&self, pub_key: &PublicKey) -> Option<Wallet> {
-        self.wallets().get(pub_key)
-    }
-}
-
-impl<'a> CurrencySchema<&'a mut Fork> {
-    pub fn wallets_mut(&mut self) -> MapIndex<&mut Fork, PublicKey, Wallet> {
-        MapIndex::new("cryptocurrency.wallets", &mut self.view)
-    }
-}
-
-transactions! {
-    // Transaction group.
-    pub CurrencyTransactions {
-        const SERVICE_ID = SERVICE_ID;
-
-        // Transaction type for creating a new wallet.
-        struct TxCreateWallet {
-            pub_key: &PublicKey,
-            name: &str,
-        }
-
-        // Transaction type for transferring tokens between two wallets.
-        struct TxTransfer {
-            from: &PublicKey,
-            to: &PublicKey,
-            amount: u64,
-            seed: u64,
-        }
-    }
-}
-
-#[derive(Debug, Fail)]
-#[repr(u8)]
-pub enum Error {
-    #[fail(display = "Wallet already exists")]
-    WalletAlreadyExists = 0,
-
-    #[fail(display = "Sender doesn't exist")]
-    SenderNotFound = 1,
-
-    #[fail(display = "Receiver doesn't exist")]
-    ReceiverNotFound = 2,
-
-    #[fail(display = "Insufficient currency amount")]
-    InsufficientCurrencyAmount = 3,
-}
-
-// Conversion between service-specific errors and the standard error type
-// that can be emitted by transactions.
-impl From<Error> for ExecutionError {
-    fn from(value: Error) -> ExecutionError {
-        let description = format!("{}", value);
-        ExecutionError::with_description(value as u8, description)
-    }
-}
-
-impl Transaction for TxCreateWallet {
-    fn verify(&self) -> bool {
-        self.verify_signature(self.pub_key())
-    }
-
-    fn execute(&self, view: &mut Fork) -> ExecutionResult {
-        let mut schema = CurrencySchema::new(view);
-        if schema.wallet(self.pub_key()).is_none() {
-            let wallet = Wallet::new(self.pub_key(), self.name(), INIT_BALANCE);
-            println!("Create the wallet: {:?}", wallet);
-            schema.wallets_mut().put(self.pub_key(), wallet);
-            Ok(())
-        } else {
-            Err(Error::WalletAlreadyExists)?
-        }
-    }
-}
-
-impl Transaction for TxTransfer {
-    fn verify(&self) -> bool {
-         (*self.from() != *self.to()) &&
-             self.verify_signature(self.from())
-    }
-
-    fn execute(&self, view: &mut Fork) -> ExecutionResult {
-        let mut schema = CurrencySchema::new(view);
-
-        let sender = match schema.wallet(self.from()) {
-            Some(val) => val,
-            None => Err(Error::SenderNotFound)?,
-        };
-
-        let receiver = match schema.wallet(self.to()) {
-            Some(val) => val,
-            None => Err(Error::ReceiverNotFound)?,
-        };
-
-        let amount = self.amount();
-        if sender.balance() >= amount {
-            let sender = sender.decrease(amount);
-            let receiver = receiver.increase(amount);
-            println!("Transfer between wallets: {:?} => {:?}", sender, receiver);
-            let mut wallets = schema.wallets_mut();
-            wallets.put(self.from(), sender);
-            wallets.put(self.to(), receiver);
-            Ok(())
-        } else {
-            Err(Error::InsufficientCurrencyAmount)?
-        }
-    }
-}
 
 struct CryptocurrencyApi;
